refactor: move mosaic to prevent circular deps
