+++ packages/lambda-xyz/src/imagery/sentinel.ts
@@ -1,4 +1,4 @@
-import { MosaicCog } from '../tiff';
+import { MosaicCog } from '../tiff.mosaic';
 
 export const MosaicSentinel = new MosaicCog('2019/new-zealand/new_zealand_sentinel_2018-19_10m/2019-12-04/', ['31'], {
     minZoom: 0,
